final release!

+++ src/main/java/com/example/pizzadelivery/controllers/PizzaRestController.java
@@ -4,16 +4,21 @@ import com.example.pizzadelivery.dtos.PizzaDto;
 import com.example.pizzadelivery.entities.Pizza;
 import com.example.pizzadelivery.mapper.PizzaMapper;
 import com.example.pizzadelivery.repositories.PizzaRepository;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.fge.jsonpatch.JsonPatch;
+import com.github.fge.jsonpatch.JsonPatchException;
 import org.modelmapper.ModelMapper;
-
-import javax.validation.Valid;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -22,6 +27,7 @@ import java.util.Optional;
 @RestController
 public class PizzaRestController {
 
+    private ObjectMapper objectMapper = new ObjectMapper();
     @Autowired
     private final PizzaRepository pizzaRepository;
     @Autowired
@@ -41,18 +47,18 @@ public class PizzaRestController {
     }
 
     @GetMapping("/pizzas")
-    List< Pizza > getAllPizzas() {
+    public List< Pizza > getAllPizzas() {
         return pizzaRepository.findAll();
     }
 
     @GetMapping("/pizzas/{id}")
     Optional< Pizza > getOnePizza(@PathVariable("id") Integer id) {
-        return pizzaRepository.findById(id);
+        return Optional.ofNullable(pizzaRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND)));
     }
 
     @GetMapping("/pizzas/search")
-    public List< Pizza > findPizzasByIngredients(@RequestParam Map< String, String > requestParams) {
-        List list = new ArrayList<Pizza>();
+    public List< Pizza > findPizzasByRequestParams(@RequestParam Map< String, String > requestParams) {
+        List list = new ArrayList< Pizza >();
         requestParams.entrySet()
                 .forEach(
                         pair -> {
@@ -72,30 +78,68 @@ public class PizzaRestController {
                             }
                         }
                 );
-       return list.stream().distinct().toList();
+        return list.stream().distinct().toList();
     }
 
     @PutMapping("/pizzas/{id}")
-    public PizzaDto updatePizza(
+    public ResponseEntity< HttpStatus > updatePizza(
             @PathVariable(value = "id") Integer id,
             @Valid @RequestBody Pizza pizzaDetails) throws ResourceNotFoundException {
-        Pizza pizza = pizzaRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Pizza not found on :: " + id));
+        try {
+            Pizza pizza = pizzaRepository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("Pizza not found on :: " + id));
 
-        pizza.setDiameter(pizzaDetails.getDiameter());
-        pizza.setIngredients(pizzaDetails.getIngredients());
-        pizza.setPrice(pizzaDetails.getPrice());
-        pizza.setType(pizzaDetails.getType());
-        return PizzaMapper.entityToDto(pizzaRepository.save(modelMapper.map(pizza, Pizza.class)));
+            pizza.setDiameter(pizzaDetails.getDiameter());
+            pizza.setIngredients(pizzaDetails.getIngredients());
+            pizza.setPrice(pizzaDetails.getPrice());
+            pizza.setType(pizzaDetails.getType());
+            PizzaMapper.entityToDto(pizzaRepository.save(modelMapper.map(pizza, Pizza.class)));
+            return ResponseEntity.status(HttpStatus.OK).build();
+        } catch (ResourceNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+    }
+
+    @PatchMapping(path = "/pizzas/{id}", consumes = "application/json-patch+json")
+    public ResponseEntity< Pizza > updatePizza(@PathVariable Integer id, @RequestBody JsonPatch jsonPatch) {
+        try {
+            Pizza pizza = pizzaRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Pizza not found on :: " + id));
+            Pizza pizzaPatched = applyPatchToPizza(jsonPatch, pizza);
+            updatePizza(pizzaPatched);
+            return ResponseEntity.ok(pizzaPatched);
+        } catch (JsonPatchException | JsonProcessingException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        } catch (ResourceNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+    }
+
+    public void updatePizza(Pizza pizza) {
+        var pizzaDto = PizzaMapper.entityToDto(pizza);
+        PizzaMapper.entityToDto(pizzaRepository.save(modelMapper.map(pizzaDto, Pizza.class)));
+    }
+
+    private Pizza applyPatchToPizza(
+            JsonPatch patch, Pizza targetPizza) throws JsonPatchException, JsonProcessingException {
+        JsonNode patched = patch.apply(objectMapper.convertValue(targetPizza, JsonNode.class));
+        return objectMapper.treeToValue(patched, Pizza.class);
     }
 
     @PostMapping("/pizzas")
-    PizzaDto saveAPizza(@RequestBody PizzaDto pizzaDto) {
+    public PizzaDto saveAPizza(@RequestBody PizzaDto pizzaDto) {
         return PizzaMapper.entityToDto(pizzaRepository.save(modelMapper.map(pizzaDto, Pizza.class)));
     }
 
     @DeleteMapping("/pizzas/{id}")
-    void deleteOnePizza(@PathVariable("id") Integer id) {
-        pizzaRepository.deleteById(id);
+    public void deletePizza(@PathVariable("id") Integer id) {
+
+        if (pizzaRepository.findById(id).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        } else
+            try {
+                pizzaRepository.deleteById(id);
+            } catch (ResponseStatusException e) {
+                e.printStackTrace();
+            }
     }
 }
